+++ src/java/util/TreeMap.d
@@ -211,7 +211,7 @@ class TreeMap : Map, SortedMap {
     public Object remove(String key){
         return remove(stringcast(key));
     }
-    public int    size(){
+    public size_t    size(){
         version(Tango){
             return map.size();
         } else { // Phobos
